design: 수강신청 관리 디자인 추가

+++ frontend/src/components/LectureEnroll/LectureEnroll.jsx
@@ -8,21 +8,39 @@ export default function LectureEnroll({ userName, enrollid, onDelete }) {
 
   const handleAccept = async (e) => {
     e.preventDefault();
+    if (!confirm('수강신청을 승인하시겠습니까?')) {
+      return;
+    }
     await lectureEnrollAccept(enrollid);
     onDelete(enrollid);
   };
 
   const handleCancel = async (e) => {
     e.preventDefault();
+    if (!confirm('수강신청을 거절하시겠습니까?')) {
+      return;
+    }
     await lectureEnrollCancel(enrollid);
     onDelete(enrollid);
   };
 
   return (
     <div className={styles.enrollLink}>
-      <p>{userName}</p>
-      <button onClick={handleAccept}>등록</button>
-      <button onClick={handleCancel}>삭제</button>
+      <span>{userName}</span>
+      <div className={styles.buttonWrapper}>
+        <button
+          onClick={handleAccept}
+          className={styles.accept}
+        >
+          등록
+        </button>
+        <button
+          onClick={handleCancel}
+          className={styles.reject}
+        >
+          삭제
+        </button>
+      </div>
     </div>
   );
 }
